requirements.txt
mysql-connector-python
fastapi
uvicorn
python-dotenv
# for form
python-multipart
# hash password
passlib[bcrypt]
argon2_cffi
# token
pyJWT
# sent request to TapPay
requests

pydantic
'pydantic[email]'